+++ CHANGELOG.md
@@ -9,7 +9,7 @@
    * ADDED: multimodal costing `auto_pedestrian` [#5780](https://github.com/valhalla/valhalla/pull/5780)
    * ADDED: `Cache-Control = public, max-age=N"` response header to `/tile` requests; also added RPATH to local debug builds to fix ldd resolving when libvalhalla is installed system-wide [#5902](https://github.com/valhalla/valhalla/pull/5902)
    * CHANGED: bump tz to 2026a [#5930](https://github.com/valhalla/valhalla/pull/5930)
-   * CHANGED: Deduplicate predicted speed profiles when updating tile [#5851](https://github.com/valhalla/valhalla/pull/5941)
+   * CHANGED: Deduplicate predicted speed profiles when updating tile [#5941](https://github.com/valhalla/valhalla/pull/5941)
 
 ## Release Date: 2026-02-19 Valhalla 3.6.3
 * **Removed**